Un-nest Cards path/rename from CardPage
